+++ src/al/alVariable.ts
@@ -1,6 +1,6 @@
 import { isUndefined } from "util";
 import { StringFunctions } from '../additional/stringFunctions';
-import { ALVarTypes } from "../additional/alVarTypes";
+import { ALVarTypes } from "./alVarTypes";
 
 export class ALVariable {
     public name: string;
@@ -10,14 +10,19 @@ export class ALVariable {
     public isTemporary: boolean = false;
     public varType: ALVarTypes | undefined;
     public varValue: string = "";
-    
+    public ignoreALPrefix: string = "";
+    public ignoreALSuffix: string = "";
+    public typeAutomaticallyDetected: boolean = false;
+
     constructor(name: string) {
         this.name = name;
     }
 
     public getVariableDeclarationString(): string {
         let declarationString = "";
-        declarationString += this.name;
+        // Only remove prefixes/suffixes if a variable suggestion name is being used
+        let variableName = this.typeAutomaticallyDetected ? this.name : StringFunctions.removePrefixAndSuffixFromVariableName(this.name, this.ignoreALPrefix, this.ignoreALSuffix)
+        declarationString += variableName;
         declarationString += ": ";
         if (this.varType) {
             declarationString += this.varType;
@@ -28,7 +33,7 @@ export class ALVariable {
         }
         if (this.objectName) {
             declarationString += " ";
-            declarationString += StringFunctions.containsSpecialChars(this.objectName)? "\"" + this.objectName + "\"" : this.objectName;
+            declarationString += StringFunctions.containsSpecialChars(this.objectName) ? "\"" + this.objectName + "\"" : this.objectName;
         }
         if (this.isTemporary) {
             declarationString += ' temporary';
@@ -36,5 +41,5 @@ export class ALVariable {
         declarationString += ';';
         return declarationString;
     }
-    
+
 }